refactor player bans
