+++ src/backend/common/data/commands/__init__.py
@@ -16,3 +16,4 @@ from .tournament_templates import *
 from .tournaments import *
 from .user_settings import *
 from .users import *
+from .player_bans import *
