+++ src/backend/common/data/models/__init__.py
@@ -13,3 +13,4 @@ from .tournament_templates import *
 from .tournaments import *
 from .user_settings import *
 from .users import *
+from .player_bans import *
